enum class must be casted to uint8 as it is does not have implicin conversion like normal enums

+++ src/GPSU/Components/MCP23017/MCP_Registers.hpp
@@ -235,9 +235,9 @@ public:
 
     // Update the specific bit in the register's value
     if (value) {
-      this->value |= (1 << field);
+      this->value |= (1 << static_cast<uint8_t>(field));
     } else {
-      this->value &= ~(1 << field);
+      this->value &= ~(1 << static_cast<uint8_t>(field));
     }
 
     // Trigger an event to notify about the change
